refactor(agents-desktop): finish settings credentials and runtime modules
Complete the Phase 4-6 extractions by moving settings persistence, Codex/API-key credential flows, runtime lifecycle, and MCP registry helpers out of main.ts.
