+++ packages/agents-desktop/src/runtime/mcp.ts
@@ -45,3 +45,64 @@ export async function handleAuthorizeUrl(deps: {
     console.warn(`[agents-desktop] OAuth flow for ${deps.server}:`, err)
   }
 }
+
+export function getMcpSnapshot(
+  snapshots: Map<string, RegistrySnapshot>,
+  serverId: string | null | undefined
+): RegistrySnapshot {
+  return (serverId ? snapshots.get(serverId) : null) ?? EMPTY_MCP_SNAPSHOT
+}
+
+export async function authorizeMcpServer(
+  runtimeEntries: Map<string, RuntimeEntry>,
+  serverId: string | null | undefined,
+  name: string
+): Promise<void> {
+  const reg = serverId
+    ? runtimeEntries.get(serverId)?.runtime?.mcpRegistry
+    : null
+  await reg?.reauthorize(name).catch((err: unknown) => {
+    console.warn(`[agents-desktop] mcp-authorize ${name}:`, err)
+  })
+}
+
+export async function reconnectMcpServer(
+  runtimeEntries: Map<string, RuntimeEntry>,
+  serverId: string | null | undefined,
+  name: string
+): Promise<void> {
+  const reg = serverId
+    ? runtimeEntries.get(serverId)?.runtime?.mcpRegistry
+    : null
+  const entry = reg?.get(name)
+  if (!reg || !entry) return
+  await reg.addServer(entry.config).catch((err: unknown) => {
+    console.warn(`[agents-desktop] mcp-reconnect ${name}:`, err)
+  })
+}
+
+export async function disableMcpServer(
+  runtimeEntries: Map<string, RuntimeEntry>,
+  serverId: string | null | undefined,
+  name: string
+): Promise<void> {
+  await runtimeEntries
+    .get(serverId ?? ``)
+    ?.runtime?.mcpRegistry?.disable(name)
+    .catch((err: unknown) => {
+      console.warn(`[agents-desktop] mcp-disable ${name}:`, err)
+    })
+}
+
+export async function enableMcpServer(
+  runtimeEntries: Map<string, RuntimeEntry>,
+  serverId: string | null | undefined,
+  name: string
+): Promise<void> {
+  await runtimeEntries
+    .get(serverId ?? ``)
+    ?.runtime?.mcpRegistry?.enable(name)
+    .catch((err: unknown) => {
+      console.warn(`[agents-desktop] mcp-enable ${name}:`, err)
+    })
+}
